% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkScore.R
\name{as_benchmark_score}
\alias{as_benchmark_score}
\title{Coercions to BenchmarkScore}
\usage{
as_benchmark_score(
  obj,
  task_id = "task_id",
  learner_id = "learner_id",
  iteration = "iteration",
  independent = TRUE,
  strip_prefix = TRUE,
  ...
)
}
\arguments{
\item{obj}{(\link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult}|\code{matrix(1)}) \cr Passed to \link{BenchmarkScore}\verb{$new()}.}

\item{task_id, learner_id, iteration, independent, strip_prefix}{See \link{BenchmarkScore}\verb{$initialize()}.}

\item{...}{\code{ANY} \cr Passed to \link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult}\verb{$score()}.}
}
\description{
Coercion methods to \link{BenchmarkScore}.
For \link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult} this is a simple wrapper around the
\link{BenchmarkScore} constructor called with \link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult}\verb{$score()}.
}
\examples{
# From a data.frame object
df = data.frame(
  tasks     = factor(rep(c("A", "B"), each = 25)),
  learners  = factor(rep(paste0("L", 1:5), each = 5)),
  iters     = 1:5,
  RMSE      = runif(50),
  MAE       = runif(50)
)
bms = as_benchmark_score(df, task_id = "tasks", learner_id = "learners", iteration = "iters")

if (requireNamespaces(c("mlr3", "rpart"))) {
  library(mlr3)
  tasks = tsks(c("boston_housing", "mtcars"))
  learners = lrns(c("regr.featureless", "regr.rpart"))
  bm = benchmark(benchmark_grid(tasks, learners, rsmp("cv", folds = 3)))

  # default measure
  as_benchmark_score(bm)

  # change measure
  as_benchmark_score(bm, measures = msrs(c("regr.rmse", "regr.mae")))
}

}
