% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkScore.R
\name{BenchmarkScore}
\alias{BenchmarkScore}
\title{Benchmark Score Result Object}
\description{
An R6 class for non-aggregated, resampling-based benchmark results.
}
\details{
This class is used as a container of benchmarking results where
multiple learners (models) have been tested against multiple tasks (datasets)
with different resamplings. The results stored are the performance scores on
the aforementioned resamplings, so multiple values per learner-task combination.

The class can either be constructed using \CRANpkg{mlr3} objects, for example
the result of \link[mlr3:BenchmarkResult]{mlr3::BenchmarkResult}\verb{$score} or via \link{as_benchmark_score}, or
by passing in a custom dataset of results. Custom datasets must include at
the very least, a factor column for learner ids, a factor column for task ids,
a numeric column for the resampling iterations (e.g. 1,2,3,...) and numeric
columns for one or more measures.
}
\examples{
# Not restricted to mlr3 objects
df = data.frame(
  tasks     = factor(rep(c("A", "B"), each = 25)),
  learners  = factor(rep(paste0("L", 1:5), each = 5)),
  iters     = 1:5,
  RMSE      = runif(50),
  MAE       = runif(50)
)
BenchmarkScore$new(df, task_id = "tasks", learner_id = "learners", iteration = "iters")
# equivalently
as_benchmark_score(df, task_id = "tasks", learner_id = "learners", iteration = "iters")

if (requireNamespaces(c("mlr3", "rpart"))) {
  library(mlr3)
  tasks = tsks(c("boston_housing", "mtcars"))
  learners = lrns(c("regr.featureless", "regr.rpart"))
  bm = benchmark(benchmark_grid(tasks, learners, rsmp("cv", folds = 3)))

  # coercion from mlr3 BenchmarkResult object
  as_benchmark_score(bm)
}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{(\link[data.table:data.table]{data.table::data.table}) \cr Resampled scores data.}

\item{\code{learners}}{\code{(character())} \cr Unique learner names.}

\item{\code{tasks}}{\code{(character())} \cr Unique task names.}

\item{\code{iterations}}{\code{(numeric())} \cr Unique resampling iterations.}

\item{\code{measures}}{\code{(character())} \cr Unique measure names.}

\item{\code{nlrns}}{\code{(integer())} \cr Number of learners.}

\item{\code{ntasks}}{\code{(integer())} \cr Number of tasks.}

\item{\code{niters}}{\verb{(integer()} \cr Number of resampling iterations.}

\item{\code{nmeas}}{\code{(integer())} \cr Number of measures.}

\item{\code{nrow}}{\code{(integer())} \cr Number of rows.}

\item{\code{col_roles}}{(\code{character()}) \cr
Column roles, currently cannot be changed after construction.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BenchmarkScore-new}{\code{BenchmarkScore$new()}}
\item \href{#method-BenchmarkScore-print}{\code{BenchmarkScore$print()}}
\item \href{#method-BenchmarkScore-summary}{\code{BenchmarkScore$summary()}}
\item \href{#method-BenchmarkScore-subset}{\code{BenchmarkScore$subset()}}
\item \href{#method-BenchmarkScore-clone}{\code{BenchmarkScore$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BenchmarkScore-new"></a>}}
\if{latex}{\out{\hypertarget{method-BenchmarkScore-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarkScore$new(
  dt,
  task_id = "task_id",
  learner_id = "learner_id",
  iteration = "iteration",
  strip_prefix = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dt}}{\code{(matrix(1))}\cr
A \code{matrix} like object coercable to \link[=data.table]{data.table::data.table},
should include column names "task_id", "learner_id", "iteration" and at
least one measure (numeric).}

\item{\code{task_id}}{(\code{factor(1)}) \cr
String specifying name of task id column.}

\item{\code{learner_id}}{(\code{factor(1)})\cr
String specifying name of learner id column.}

\item{\code{iteration}}{(\code{character(1)})\cr
String specifying name of iteration id column.}

\item{\code{strip_prefix}}{(\code{logical(1)}) \cr
If \code{TRUE} (default) then mlr prefixes, e.g. \code{regr.}, \code{classif.}, are automatically
stripped from the \code{learner_id} and other columns (i.e. representing measure ids).}

\item{\code{...}}{\code{ANY} \cr
Additional arguments, currently unused.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BenchmarkScore-print"></a>}}
\if{latex}{\out{\hypertarget{method-BenchmarkScore-print}{}}}
\subsection{Method \code{print()}}{
Prints the internal data via \link[data.table:print.data.table]{data.table::print.data.table}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarkScore$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{\code{ANY} \cr Passed to \link[data.table:print.data.table]{data.table::print.data.table}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BenchmarkScore-summary"></a>}}
\if{latex}{\out{\hypertarget{method-BenchmarkScore-summary}{}}}
\subsection{Method \code{summary()}}{
Prints the internal data via \link[data.table:print.data.table]{data.table::print.data.table}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarkScore$summary(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{\code{ANY} \cr Passed to \link[data.table:print.data.table]{data.table::print.data.table}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BenchmarkScore-subset"></a>}}
\if{latex}{\out{\hypertarget{method-BenchmarkScore-subset}{}}}
\subsection{Method \code{subset()}}{
Subsets the data by given tasks, learners and iterations.
Returns data as \link[data.table:data.table]{data.table::data.table}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarkScore$subset(tasks = NULL, learners = NULL, iterations = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tasks}}{(\code{character()}) \cr
Task(s) to subset the data by.}

\item{\code{learners}}{(\code{character()}) \cr
Learner(s) to subset the data by.}

\item{\code{iterations}}{(\code{numeric()}) \cr
Resampling iteration(s) to subset the data by.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BenchmarkScore-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BenchmarkScore-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BenchmarkScore$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
