% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3benchmark-package}
\alias{mlr3benchmark}
\alias{mlr3benchmark-package}
\title{mlr3benchmark: Analysis and Visualisation of Benchmark Experiments}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Implements methods for post-hoc analysis and visualisation of benchmark experiments, for 'mlr3' and beyond.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3benchmark.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3benchmark}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3benchmark/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Fischer \email{sebf.fischer@gmail.com} (\href{https://orcid.org/0000-0002-9609-3197}{ORCID}) [contributor]

Authors:
\itemize{
  \item Sonabend Raphael \email{raphaelsonabend@gmail.com} (\href{https://orcid.org/0000-0001-9225-4654}{ORCID})
  \item Florian Pfisterer \email{pfistererf@googlemail.com} (\href{https://orcid.org/0000-0001-8867-762X}{ORCID})
}

Other contributors:
\itemize{
  \item Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID}) [contributor]
  \item Bernd Bischl \email{bernd_bischl@gmx.net} (\href{https://orcid.org/0000-0001-6002-6980}{ORCID}) [contributor]
}

}
