% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkAggr.R
\name{as.BenchmarkAggr}
\alias{as.BenchmarkAggr}
\title{Coercions to BenchmarkAggr}
\usage{
as.BenchmarkAggr(obj, ...)
}
\arguments{
\item{obj}{\verb{([mlr3::BenchmarkResult]|matrix(1))} \cr Passed to \verb{[BenchmarkAggr]$new()}.}

\item{...}{\code{ANY} \cr Passed to \verb{[BenchmarkResult]$aggregate()}.}
}
\description{
Coercion methods to \link{BenchmarkAggr}
}
