% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkAggr.R
\name{as.BenchmarkAggr}
\alias{as.BenchmarkAggr}
\title{Coercion to BenchmarkAggr}
\usage{
as.BenchmarkAggr(...)
}
\arguments{
\item{...}{(any)\cr
Passed to \code{\link[=as_benchmark_aggr]{as_benchmark_aggr()}}.}
}
\description{
Coercion method to \link{BenchmarkAggr}.
Only there for backwards compatability. Use \code{\link[=as_benchmark_aggr]{as_benchmark_aggr()}} instead.
}
