% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkAggr.R
\name{autoplot.BenchmarkAggr}
\alias{autoplot.BenchmarkAggr}
\title{Plots for BenchmarkAggr}
\usage{
\method{autoplot}{BenchmarkAggr}(obj, type = c("mean", "cd"), meas = NULL, level = 0.95, ...)
}
\arguments{
\item{obj}{\link{BenchmarkAggr}}

\item{type}{\code{(character(1))} \cr Type of plot, see description.}

\item{meas}{\code{(character(1))} \cr Measure to plot, should be in \code{obj$measures}, can be \code{NULL} if
only one measure is in \code{obj}.}

\item{level}{\code{(numeric(1))} \cr Confidence level for error bars for \code{type = "mean"}}

\item{...}{\code{ANY} \cr Additional arguments passed to \verb{[BenchmarkAggr]$crit_differences}.}
}
\description{
Generates plots for \link{BenchmarkAggr}, depending on argument \code{type}:
\itemize{
\item \code{"mean"} (default): Assumes there are at least two independent tasks. Plots the sample mean
of the measure for all learners with error bars computed with the standard error of the mean.
\item \code{"cd"}: Critical difference plots (Demsar, 2006), uses the \verb{$crit_differences} method from
\link{BenchmarkAggr}. If a baseline is selected for the Bonferroni-Dunn test, the critical difference
interval will be positioned around the baseline. If not, the best performing algorithm will be
chosen as baseline. Learners are drawn on the y-axis according to their average rank.
For \code{test = "nemenyi"} a bar is drawn, connecting all groups of not
significantly different learners. For \code{test = "bd"} an interval is drawn around the algorithm
selected as a baseline and any learner within this interval is not significantly different
from the baseline.
}
}
\references{
Janez Demsar, Statistical Comparisons of Classifiers over
Multiple Data Sets, JMLR, 2006
}
