% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3benchmark-package}
\alias{mlr3benchmark}
\alias{mlr3benchmark-package}
\title{mlr3benchmark: Benchmarking analysis for 'mlr3'}
\description{
Implements methods for post-hoc analysis and
    visualisation of benchmark objects in 'mlr3'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3benchmark.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3benchmark}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3benchmark/issues}
}

}
\author{
\strong{Maintainer}: Sonabend Raphael \email{raphael.sonabend.15@ucl.acuk} (-13880)

}
