% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3benchmark-package}
\alias{mlr3benchmark}
\alias{mlr3benchmark-package}
\title{mlr3benchmark: Analysis and Visualisation of Benchmark Experiments}
\description{
Implements methods for post-hoc analysis and
    visualisation of benchmark experiments, for 'mlr3' and beyond.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3benchmark.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3benchmark}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3benchmark/issues}
}

}
\author{
\strong{Maintainer}: Sonabend Raphael \email{raphael.sonabend.15@ucl.ac.uk} (\href{https://orcid.org/0000-0001-9225-4654}{ORCID})

Authors:
\itemize{
  \item Florian Pfisterer \email{pfistererf@googlemail.com} (\href{https://orcid.org/0000-0001-8867-762X}{ORCID})
}

}
