% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{requireNamespaces}
\alias{requireNamespaces}
\title{Helper Vectorizing requireNamespace}
\usage{
requireNamespaces(x)
}
\arguments{
\item{x}{Packages to check.}
}
\description{
Internal helper function for documentation.
}
